chore(deps): update `typescript-eslint` to v8 (#2767)
Updated `typescript-eslint` package to v8. Some of the rules have moved
from `@typescript-eslint/eslint-plugin` to `@stylistic/eslint-plugin`.
Therefore added that as a dependency. Also updated
`eslint-config-streamr-ts` to v5 as it was earlier version was not
compatible with `typescript-eslint` v8.

Added `eslint-disable-next-line` annotations for all new warnings. We
should go through these cases in this PR or in a follow-up PR.

| Rule                                             | Occurrences |
|:-------------------------------------------------|:------------|
| @typescript-eslint/prefer-promise-reject-errors  | 9           |
| @typescript-eslint/only-throw-error              | 4           |
| @typescript-eslint/no-empty-object-type          | 2           |
| @typescript-eslint/no-unsafe-function-type       | 1           |
| @typescript-eslint/no-array-delete               | 1           |