chore(sdk): Use `MaxUint256` in permissions (#2778)
Changed `setPermissions()` to use `MaxUint256` instead of the signed
version (`MaxInt256`) when it writes a permission record to the
contract.

This change doesn't change functionality in practice, as both values are
enormously big numbers. But changed this for consistency: e.g. if we
grant the same permission with `grantPermission`, the contract correctly
writes `MaxUint256` to `publishExpiration` and `subscribeExpiration`.