chore(sdk): Remove `webrtcNewConnectionTimeout` config option (#2776)
Removed obsole `webrtcNewConnectionTimeout` config option. It is
obsolete, as the corresponding DHT config option was removed (see
https://github.com/streamr-dev/network/pull/2559).

There was a mismatch between `Config` and `config.schema.json`: in the
schema the property had a deprecated name of
`newWebrtcConnectionTimeout`. The option was renamed in commit 6c8803f.
The mismatch means that it has not been possible to use renamed config
option in the client as it would have failed the schema validation.