chore(sdk): Remove node id config option (#2777)
Removed obsolete `NetworkNodeConfig#id` config option. The
`NetworkNodeConfig` interface corresponds the
`ContentDeliveryManagerOptions` interface in `trackerless-network`, and
there is no `id` field in that interface. Therefore the config option
doesn't configure anything and is obsolete.

## Background

The description of the removed config option was "the Ethereum address
of the node". Therefore it is likely that the config option is from
pre-`trackerless-network` era. In that implementation node ids were
Ethereum addresses.

## Other changes

Fixed config injection problem in `OperatorPlugin` test. When the
"invalid configuration" test case modified the config object returned by
`formConfig`, it modified also the `TEST_CONFIG` object (because
`TEST_CONFIG` object properties are injected by reference in
`formConfig`). That caused another test in the `OperatorPlugin` to fail.
- Alternatively we could have done the the cloning in `formConfig`. But
as cloning is needed only in this one test case, maybe better to do it
locally.