feat(network): NET-1346 avoid routing through proxy connections (#2801)
## Summary

Added new functionality in the connection manager to set the manager in
private mode. In private mode the node will inform all peers that the
connection is private. This way the peers will know not to use the
connections during routing.

This fixes a known issue with routing where routing is attempted through
pure proxy client nodes.

Allowing private connections has to be explicitly allowed. This is
required to not allow free riding. Private connections are allowed if
the `acceptProxyConnections` flag is set true in the SDK. There is
currently no need to have an option to only set the
`acceptIncomingPrivateConnections` as true since the feature is only
used for proxies.