feat(utils): `merge()` does deep merge (#2812)
Changed the `merge()` utility function to perform a deep merge instead
of a flat merge. In most cases, it doesn't matter whether we use flat or
deep merging, so the flat behavior may not have been explicitly
required. When merging objects, we could implicitly assume that the
merge is deep. So maybe this is better this way?

This is part of a preparation for NET-1363, where a deep merge is
needed.

Also simplified `applyEnvironmentDefaults()`.

## Usage in Stream#update

The `Streamr#update` is one of the callers of `merge()`. The API
documentation says it "merges metadata with the existing metadata.". It
is not clear whether that means a deep or a flat merge?

We'll harmonize the `Stream#update` method soonish so that it matches
the functionality of `StreamrClient#updateStream` (NET-1364). That's why
we don't change the behavior in this PR. If we'd keep the merging
functionality, it should most likely do deep merge.