test(network): fix test flakyness `end-to-end/websocket-full-node-network.test.ts`  (#2833)
## Summary

After #2827 the increased cpu load is causing problems in the websocket
test. This seems to be caused by a known issue with local connections
with the `ws` library especially since it does not really appear in the
browser. This issue should not exist in production but it probably would
make sense to investigate it. (Possibly related to NET-1349)