refactor(DHT: NET-1366 move PeerDescriptor out of DhtRpc, no longer import DhtRpc to higher layers (#2820)
## Summary

Move PeerDescriptor out of DhtRpc to reduce the amount of generated code
in higher level packages. DhtRpc is no longer imported in the higher
layer packages. Instead only the PeerDescriptor is imported. This means
that any imports pointing to the DhtRpc had to be changed

PR is large due the number of generated files being deleted. See commits
to see what changes in each package.