feat(sdk)!: `StreamMetadata#partitions` is nullable (#2825)
**This is a breaking change** as this changes the API

Changed `StreamMetadata` field type: the `partitions` field now
nullable.

We can't ensure that the data contains some values (or even that
metadata is a JSON). Therefore it is better that the type system doesn't
assert anything about the content of the metadata.

Also added `Stream#getPartititionCount()` method. It reads the info from
the metadata and defaults to 1 if there is no data available. It also
validates that the partition count is valid (between 1 and 100).

Also added `DEFAULT_PARTITITION_COUNT` constant to `@streamr/utils`.

## Future improvements

Maybe the best type for metadata would be just `Record<string, any>`?
- also maybe the `Stream` constructor and `Stream#parseMetadata`
shouldn't inject any default value for the `partitions` field