feat(sdk)!: `Stream#addToStorageNode` wait options (#2810)
**This is a breaking change**

Changed the behavior of `Stream#addToStorageNode`: now it doesn't wait
for acknowledgment by default. If waiting is needed, the method should
be called like this:
```ts
await stream.addToStorageNode(address, { wait: true })
```

## Background

There are two methods for adding a stream to storage node:
`Stream#addToStorageNode` and `StreamrClient#addStreamToStorageNode`.
These should work similarly, but before this PR the `Stream` method
always waited for ack, and `StreamrClient` never did that. That was not
due to a bug, but an inconsistent API design.

This PR is part of NET-1364 in which we harmonize the
`Stream`/`StreamrClient` API.

## Open question: API design

Is this the best API? There are three main options:
- opts is optional, and we don’t wait by default
- opts is optional, and we do wait by default
- opts is required

When interacting with contracts, we usually don't include wait
parameters - we wait for the transaction to be mined and resolve after
that. The `{ wait: true }` option is needed in tests, where we publish
to the stream right after making the assignment call.

In production, these actions are often split into two steps: first,
setting up the environment (creating a stream, setting permissions,
assigning a storage node), then starting the application that publishes
real-time data as it’s generated.

## Future improvements

- Update `StreamrClient#addToStorageNode` to match the
`Stream#addToStorageNode`