feat(sdk): User id as `HexString` (#2774)
## Summary

User id items are stored as arbitrary length `string` instead of
`EthereumAddress`.

Renamed `getAddess()` in `StreamrClient` to `getUserId()`. **This is a
breaking change.**

This is a preparation PR for
https://github.com/streamr-dev/network/pull/2780. This PR doesn't
doesn't contain any changes to contract / The Graph interaction and
therefore any user ids which are not exactly 20 bytes long don't work
yet. (We should merge this PR to `main` just before that PR, or
implement the contract interactions in a PR which has this PR as a base
branch.)

## Changes

The `UserID` helper type if now a a separate `BrandedString`.
Additionally there is a new helper type `UserIDRaw`, and helper
functions `toUserId` and `toUserIdRaw`. There is also the `HexString`
helper type which is an alias to `string`.

In the API all user id inputs and outputs are `HexString` values. This
is not a breaking change as earlier the inputs and outputs were either
`string` or `EthereumAddress` values. (The `EthereumAddress` is a
branded string, and it was used only in outputs, i.e. in the API all of
these were just `string` values.)

In `Message`:
- field type of `publisherId`

In `StreamrClient`:
- return type of `getUserId()`,
- field type of `UserPermissionQuery#user`
  - it is parameter type in `hasPermission()`
- field type of `UserPermissionAssignment#user`
- it is parameter/return type in `getPermissions()`,
`grantPermissions()`, `revokePermissions()` and `setPermissions()`
- return type of `getStreamPublishers()` and `getStreamSubscribers()`
- parametrer type of `isStreamPublisher()`, `isStreamSubscriber()`
- field type of `SearchStreamsPermissionFilter#user`
  - it is a parameter type in `searchStreams()`
- changed type of `ResendFromOptions#publisherId` and
`ResendRangeOptions#publisherId`
  - these are parameter types in `resend()`
- parameter type of `addEncryptionKey()`

In `Stream`:
- corresponding changes in `hasPermission()`, `getPermissions()`,
`grantPermissions()`, `revokePermissions()` and `setPermissions()`

### Internal

In `StreamMessage` the `publisherId` field type is `UserID` (i.e. the
branded string) as we don't consider that class to be a part part of the
public API. Also the related user id types in
`MessageID`/`GroupKeyRequest`/`GroupKeyResponse` are `UserID` strings.

## Future improvements

- Change all `EthereumAddress` outputs to `HexString` outputs (e.g.
`getStorageNodes()`) to unify the API
- Could rename `user` field of `UserPermissionAssignment#user` and
`SearchStreamsPermissionFilter#user` to `userId` to unify the API (we
have typically the ID suffix in these, e.g. `streamId` and
`publisherId`)
- Could refactor some test so that we use `randomUserId` instead of
`fastWallet` / `ether`'s `Wallet`