feat(sdk): Validate partition count when it is used (#2854)
## Summary

Do not validate partition count when a `Stream` instance is created
(e.g. by `Stream#getStream()`). Instead validate it when the value is
used.

## Background

This is a follow-up PR for a PR which changed stream data from a typed
structure to untyped `Record<string, unknown>`
(https://github.com/streamr-dev/network/pull/2845).

In a previous PR (https://github.com/streamr-dev/network/pull/2825) we
added an accessor `Stream#getPartitionCount()`. It validates the value
before it is used.

After that change the partition validation was in two separate places:
in `Stream#getPartitionCount()` and in `Stream#parseStream()`. It is
better to have the validation only in one place. We could have kept the
validation only in `parseStream()`, but as metadata is now untyped, it
maybe makes more sense to keep it only in `getPartitionCount()`.

There is no fundamental reason why we shouldn't allow `StreamrClient` to
process streams which have invalid metadata. Users could use the client
e.g. for fixing invalid metadata:
```ts
(await client.getStream(...)).update({ partitions: 10 })
```

## Changes

The internal `Stream#parseStream()` accepts all metadata values. If the
input string is not JSON, we parse it as an empty object.

## Implications

If a stream has invalid metadata, these functionalities are now
different:
- `StreamrClient#getStream()` succeeds, i.e. it returns a `Stream`
instance
- the stream is not omitted from `StreamrClient#searchStreams()`
- the stream is not omitted from `StreamrClient#getStoredStreams()`
- `StreamrClientEvents#streamCreated` event is emitted for the stream

## Statistic
- There are 4911 streams in the production environment
- ~1% of the streams have empty `{}` object as metadata, most of those
are test streams
- ~0.5% of the streams have non-JSON metadata, all of those seem to be
test streams
- <0.1% of the streams have JSON metadata without the `partitions` field

## Other changes

- Changed `getStreamPartitionCount()` to throw `StreamrClientError`
instead of the generic `Error` if partition count is invalid

## Future improvements

- Maybe `StreamrClient#createStream` and `StreamrClient#updateStream()`
should validate the partition count