refactor(sdk): `StreamRegistry#getStreamMetadata()`  (#2856)
Small refactoring to `StreamRegistry`:
- added `getStreamMetadata()` method
- removed `getStream()` method

The `getStream()` method just fetched the metadata and created a new
instance of `Stream` class. It is better that `StreamRegistry` operates
only on contract concepts, e.g. metadata, and not high-level concepts
like the `Stream` class.

This we were also able to encapsulate the handling of metadata caching
into `StreamRegistry`. Now `StreamRegistry#updateStreamMetadata()`
clears the cache internally.
- This change also fixes a small issue. Previously `Stream#update()`
cleared the cache explicitly, but `StreamrClient#updateStream()` did not
do that. Now cache is cleared in both execution paths.

## Other changes

- Moved `StreamMetadata` functionality to a separate file
- Renamed `StreamRegistry#updateStream()` to `updateStreamMetadata()`

## Future improvements

- As metadata is now cached, the `Stream` object could be stateless (the
`metadata` field is only stateful field in `Stream`)
- Could continue this refactoring by moving all usages of
`StreamFactory` in `StreamRegistry` to the caller (i.e. `StreamrClient`)