refactor(sdk): Create `Stream` instances outside of `StreamRegistry`/`StreamStorageRegistry` (#2857)
Moved `Stream` instance creation from `StreamRegistry` and
`StreamStorageRegistry` to `StreamrClient`. This is a follow-up PR to
https://github.com/streamr-dev/network/pull/2856 in which similar
refactoring was done for `metadata`.

## Other changes

The class dependency graph changed as `StreamFactory` is no longer a
dependency in registries. That caused circular dependency problem in DI.
To fix the problem the `delay()` annotation was added to the `Resends`.
```
npm run test-unit

Cannot inject the dependency "messagePipelineFactory" at position #1 of "Resends" constructor. Reason:
Attempted to construct an undefined constructor. Could mean a circular dependency problem. Try using `delay` function.
```