chore: Eslint v9 (#2849)
## Changes

Converted the configuration file to flat config and unified to config
file style. Otherwise same rules are earlier, except:
- some new rules are now inherited from the updated
`eslint-config-streamr-ts`
  - removed locally defined rules if exactly same as inherited
- converted `cli-tools` logging statements to use `console.info` instead
of `console.log` and removed the `no-console` rule from that package

Fixed lint errors and removed obsolete `eslint-disable` directives. Also
fixed a type in `Operator.test.ts`.

Added `eslint-import-resolver-typescript` dependency. It is needed by
`eslint-plugin-import` (see
https://github.com/import-js/eslint-plugin-import?tab=readme-ov-file#typescript).

## Other PRs

Temporarily disabled linting of browser classes in commit 876423f. That
is needed until we merge
https://github.com/streamr-dev/network/pull/2848 to `main`. We should
revert the commit either in this PR or in that PR (whichever is landed
to `main` later).