feat(sdk)!: Remove `StreamrClient#findOperators()` (#2876)
**This is a breaking change as this changes the API**

Removed the method from public API by annotating it as `@internal`. 

The API of this method and some other related methods may change soon
(NET-1374).

## Background

The method signature use internal `StreamID` type. If we re-include the
method to public API that parameter type should be changed to
`streamIdOrPath: string`.

Also the magic numbers (`maxQueryResults`, `maxHeartbeatAgeHours`)
should be configurable, or the method name should reflect the fact that
it finds only a subset of operators? Note also that query currently
excludes non-TLS operator nodes. We could either add a configuration
option to handle that exclusion or rename the method (e.g.
`findConnectableOperators()`).