chore: Add `@typescript-eslint/no-confusing-void-expression` rule (#2910)
Added new `eslint` rule:
https://typescript-eslint.io/rules/no-confusing-void-expression/

This rule disallows e.g. using `return foobar()` when `foobar()` method
returns `void`. It is better to just call the method and do explicit
return separately, if needed. (Quite typically the `return` is not
needed at all as the line is the last line of the method.)

Simplified `multiple-publisher-plugins.test.ts` so that it uses
`Promise<void>` instead of `Promise<unknown>` and therefore doesn't need
any special handling related to this rule.