feat(sdk): `Mapping` cache constraints (#2901)
Added `maxSize` and `maxAge` options to the `Mapping` class. These are
needed in follow-up PR
(https://github.com/streamr-dev/network/pull/2903).

Added two builder functions: `createCacheMap()` and `createLazyMap()`.
This way the user of the class needs to decide if the Mapping if is a
cache or just a lazy loading map. For caches it is required to configure
`maxSize`.

## Type safety

The constructor of `Mapping` is annotated as `@internal` so that in
encourages the usage of the builder functions. If we make it private,
the builder functions needs to have access to it somehow. Could e.g. be
static methods. Or we could export `Mapping` as a type.

## Testing

In unit tests we test the basic functionality for the lazy map, and use
cache map only test the size limitations. This is maybe ok as in the
`Mapping` the differences are abstracted out to the `delegate` (which is
either a LRU cache or plain `Map`).

## Future improvements

- Use `Mapping` instead of `CacheMap` in
`StreamRegistry`/`StreamStorageRegistry` and remove the `CacheMap` class