fix(network): Revert backwards incompatible PR #2820 (#2922)
Reverting PR #2820 due to backwards incompatibility with production. The
typeUrl of the PeerDescriptors in the Any protobuf message has changes
from `dht.PeerDescriptor` to `peerDescriptor.PeerDescriptor` in this
change. This makes it impossible to parse the stream entry points stored
stored in the DHT causing stream joining to not work.

The clean up of generated protobuf code will be implemented in #2842