refactor(sdk): Improve `InternalPermissionQuery` type safety (#2925)
The type of `streamId` field is now `StreamID` instead of `string`.

Also remove obsolete `toInternalPermissionQuery()` call in
`Stream#hasPermissions()`. There is no need to convert the query as the
we call the client method which does the same conversion.

Also fixed method signatures in `FakeStreamRegistry` so that they
matched `StreamRegistry` methods.