refactor(sdk): Use `Mapping` instead of `CachingMap` (#2903)
Use `Mapping` cache instead of `CachingMap` in `StreamRegistry` and
`StreamStorageRegistry`. After these modifications there are no usages
for `CacheMap` in `sdk`, and we can remove it (in a follow-up PR).

## Testing

The implementations of `Mapping` and `CachingMap` are very different,
but the functionality should be almost/exactly same. Double-check that
there are no corner case in which behavior is significantly different.
Checked that CachingMap tests pass also with the new implementation (see
temporary commit
[e5bfdcf](https://github.com/streamr-dev/network/pull/2903/commits/e5bfdcfc94db2af5c0fc46c22261f4022a8473db)).
The same functionality was already covered by `Mapping.test.ts`, so
there was no need keep any `CachingMap.test.ts` test cases for the
`Mapping` class.

## Other changes

- Changed argument type of `Mapping#get()` and `Mapping#valueFactory()`,
and return type of `Mapping#values()`

## Future improvements

- Remove `CachingMap`