test(sdk): Custom matchers accept message substrings (#2941)
Modified the functionality of `toThrowStreamrClientError()` and
`toEqualStreamrClientError()` so that the functions handle error
messages similarly to Jest's `toThrow()` function. In Jest the the error
message must match exactly if the provider argument is an `Error`
object, but a substring match is accepted if the provider argument is a
`string`. See https://jestjs.io/docs/expect#tothrowerror.

Before this PR the custom matchers required an exact match in both
cases.

Also modified some tests the `toEqualStreamrClientError()` function and
improved test coverage in `customMatchers.test.ts`.