feat(sdk)!: Export `AbstractSigner` instead of `Signer` (#2939)
Changed SDK to export `ether`'s `AbstractSigner` type instead of
`Signer` type.

## Background

The type is exported because it is part of our public API. It is used
in`SignerWithProvider` type (which is used by
`StreamrClient#getSigner()`). In PR
https://github.com/streamr-dev/network/pull/2506 we changed
`SignerWithProvider` to use `AbstractSigner` instead of `Signer`. We
should have changed the export in that PR.