feat(docs): Hide internal methods from API docs (#2945)
Added `@hidden` TypeDoc annotation for internal methods:
- `getNode()`
- `getConfig()`

This way these methods are not included in the API docs. The internal
`StreamrClient#getOperator()` was already annotated like this.

Removed redundant `@internal` annotations from the `NetworkNodeStub` as
it is not used by any publicly available method anymore:
- also modified `NetworkNodeFacade#getNode()` to return
`Promise<Omit<NetworkNodeStub, 'start' | 'stop'>>` as a consequence of
removing the `@internal` annotations
- also removed unnecessary `@internal` annotation from
`NetworkNodeFacade#setProxies()`

## Note

It would be possible annotate those method with `@internal`. But then
the methods wouldn't be included to the generated `.d.ts` files (as we
use `"stripInternal": true` in `tsconfig.node.json`). The type
definitions are needed by `node` and `cli-tools` packages which use
these internal methods.