refactor: remove generated protobuf files from the repo (#2842)
## Summary

Removing generated protobuf files from the public repo

## Future improvements

- What should we do with the proto-rpc package's generate protobuf code?
Currently the solution requires generateing the code for eslint, check,
and test separately. We don't want to include the building of the test
files in pre-build.
- Could look into pruning all of the bootstrapped packages in the
Dockerfile. Based on a quick test the size of the image is the same if
only the node is pruned or if all sub packages are pruned separately.