test: move `generateWalletWithGasAndTokens()` to `test-utils` (#2963)
Moved `generateWalletWithGasAndTokens()` test utility from `sdk` package
to the `test-utils` package.

## Small changes

- The `setupOperatorContract()` function in `sdk` now takes
`generateWalletWithGasAndTokens` function as a parameter. In theory the
`setupOperatorContract()` function is "production code", so it can't
have a direct dependency to test code, i.e. to the
`generateWalletWithGasAndTokens()` implementation. We'll refactor
`setupOperatorContract()` and related contract utilities in a follow-up
PR to generalize and simplify the usage. After that refactoring this
hack is no longer needed.
- When the implementation was in `sdk`, the test token ABI was read from
a local `DATAv2Abi.json`. Now it is hardcoded into
`getTestTokenContract()` as we need a small fraction of the ABI. After
https://linear.app/streamr/issue/ETH-799/collect-all-contracts-into-streamrcontracts
we can read the ABI from network-contracts in both places.