Switch to using helsinki.fi instead of suomi.fi in GeoIpLocation tests (#2974)
* The coodrinates of suomi.fi had apparently changed in the geoip
databases, which caused the GeoIpLocation tests to fail
* This pull request fixes the problem by using helsinki.fi as the
address whose location is checked in the tests instead. Helsinki.fi is
the domain of the university of Helsinki, and its coordinates are
unlikely to change in the near future as the university has its own IT
department and the domain is not hosted at third-party cloud providers.