feat(network): limit pings in layer1 NET-1384 (#2979)
## Summary

Limit the number of neighbors to ping when they are added as neighbors.
This fixes a bad looping behaviour during joins to layer1 DHTs

## Side effects

While all nodes during the start of the joins are pinged, after 16
neighbors have been added there is a chance that new neighbors are
offline. This will be resolved eventually with the use of periodic pings
to all neighbors. However it can lead to issues with layer2 graph
generation if multiple nodes leave the network ungracefully in a small
time frame.