feat(cli-tools): [NET-1398] Add internal operator and sponsorship commands (#2926)
Added these internal commands:
- `sponsorship-sponsor`
- `operator-delegate`
- `operator-undelegate`
- `operator-stake`
- `operator-unstake`

## API

The token amounts are given in full DATA tokens, not in DATA token weis.
This is maybe a good approach for end-user tools like this.

## Other changes

Ractored some `operatorContractUtils` method signatures: user
authentication is passed as `SignerWithProvider` instead of `Wallet`, so
that `client.getSigner()` can be used as an argument. These are internal
methods, i.e. this is not a breaking change.

## Future improvements

These commands could be normal CLI tools commands, i.e. just not for the
internal use. That would be possible if we add support for `DataV2`
token to the SDK. Currently the only token contract available for
`operatorContractUtils` is the test token contract. (Note that most
likely there is no need to have both the `DataV2` and `TestToken`, as
`DataV2` can be used also for tests.)