build(deps): [NET-1399] Bump `electron` (#2977)
Upgraded `electron` from v31 to v34.

## Background

The release notes of
[v31.7.7](https://github.com/electron/electron/releases/tag/v31.7.7):
"Electron 31.x.y has reached end-of-support as per the project's support
policy. Developers and applications are encouraged to upgrade to a newer
version of Electron."

## Related changes

After v32 upgrade the `npm run versions` started to fail with this
error:
```
(node:2310) ExperimentalWarning: WASI is an experimental feature and might change at any time
(Use `node --trace-warnings ...` to show where the warning was created)
SyntaxError: Unexpected token '(', "(node:2310"... is not valid JSON
    at JSON.parse (<anonymous>)
    at file:///home/runner/work/network/network/show-versions.mjs:8:23
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
Error: Process completed with exit code 1.
```

The `show-versions.mjs` script was modified to avoid the error. It is
unclear how the `electron` v32 affected this script. The wasi features
were added to the script already before this PR (in `nx`
[v19.5.0](https://github.com/nrwl/nx/releases/tag/19.5.0), which is a
dependency of `lerna`).